# backend/openapi.yaml
...
openapi: 3.0.3
info:
  title: CarbonManager On-Chain API
  version: 1.0.0
servers:
  - url: http://localhost:5001
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
paths:
...
  /product_types:
    get:
      summary: Get all available product types from this organization 
      tags: [Product types]
      security:
        - BearerAuth: []
      responses:
        "200":
          description: List of product types
          content:
            application/json:
              schema:
                type: object
                properties:
                  product_types:
                    type: array
                    items:
                      type: object
                      properties:
                        product_type_id:
                          type: string
                          example: PT1
                        product_type_name:
                          type: string
                          example: Green Tea 2025
        "500":
          description: Server error
    post:
      summary: Create a new product type
      tags: [Product types]
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                product_type_name:
                  type: string
                  example: Green Tea 2025 
      responses:
        "201":
          description: Product type created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  product_type_id:
                    type: string
                    example: PT1
                  product_type_name:
                    type: string
                    example: Green Tea 2025
                  organization_id: 
                    type: string
                    example: ORG4
                  organization_name:
                    type: string
                    example: Lee's Tea
                  created_at: 
                    type: string
                    example: 2024-01-01T12:00:00Z
                  updated_at: 
                    type: string
                    example: 2024-01-01T12:00:00Z
                  order_id:
                    type: int
                    example: 1

        "400":
          description: Invalid input
        "401":
          description: Unauthorized
        "409":
          description: Product type already exists
  /product_types/{product_type_id}:
      get:
        summary: Get a product type by ID
        tags: [Product types]
        security:
          - BearerAuth: []
        parameters:
          - in: path
            name: product_type_id
            required: true
            schema: { type: string, minimum: 1 }
        responses:
          "200":
            description: Product type details
            content:
              application/json:
                schema:
                  type: object
                  properties:
                    product_type_id:
                      type: string
                      example: PT1
                    product_type_name:
                      type: string
                      example: Green Tea 2025
                    organization_id:
                      type: string
                      example: ORG4
                    organization_name:
                      type: string
                      example: Lee's Tea
                    created_at: 
                      type: string
                      example: 2024-01-01T12:00:00Z
                    updated_at:
                      type: string
                      example: 2024-01-01T12:00:00Z
          "404":
            description: Product type not found
      put:
        summary: Update a product type by ID
        tags: [Product types]
        security:
          - BearerAuth: []        
        parameters:
          - in: path
            name: product_type_id
            required: true
            schema: { type: string, minimum: 1 }
        requestBody:      
          required: true
          content:
            application/json:
              schema:
                type: object
                properties:
                  name:
                    type: string
                    example: Updates Green Tea 2025
        responses:
          "200":
            description: Product type updated successfully
          "400":
            description: Invalid input
          "401":
            description: Unauthorized
          "404":
            description: Product type not found
      delete:
        summary: Delete a product type by ID
        tags: [Product types]
        security:
          - BearerAuth: []
        parameters:
          - in: path
            name: product_type_id
            required: true        
            schema: { type: string, minimum: 1 }
        responses:
          "200":
            description: Product type deleted successfully
            content:
              application/json:
                schema:
                  type: object
                  properties:
                    status:
                      type: string
                    product_type_id:
                      type: string
                    product_type_name:
                      type: string
                    order_id:
                      type: integer
                    
          "401":
            description: Unauthorized     
          "404":
            description: Product type not found
  /product_types/{product_type_id}/products:
    get:
      summary: Get all products under a product type
      tags: [Products]
      security:
        - BearerAuth: []
      parameters:
        - name: product_type_id
          in: path
          required: true
          schema: { type: string, minimum: 1 }
      responses:
        "200":
          description: List of products
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    product_id:
                      type: string
                    product_name:
                      type: string
    post:
      summary: Create a new product under a product type
      tags: [Products]
      security:
        - BearerAuth: []
      parameters:
        - name: product_type_id
          in: path
          required: true
          schema: { type: integer, minimum: 1 }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  example: Batch #12345
      responses:
        "201":
          description: Product created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  product_id:
                    type: string
                    example: PRD1
                  product_name:
                    type: string
                    example: Green Tea Batch 
        "400":
          description: Invalid input
        "401":
          description: Unauthorized
        "404":
          description: Product type not found
        "409":
          description: Product already exists
  /products/{product_id}:
    get:
      summary: Get a product by ID
      tags: [Products]
      security:
        - BearerAuth: []
      parameters:
        - name: product_type_id
          in: path
          required: true
          schema: { type: string, minimum: 1 }
        - name: product_id
          in: path
          required: true
          schema: { type: string, minimum: 1 }     
      responses:
        "200":
          description: Product details
          content:
            application/json:
              schema:   
                type: object
                properties:
                  product_id:
                    type: integer
                    example: PD1
                  product_name:
                    type: string
                    example: Green Tea Batch
                  serial_number:
                    type: string
                  total_emission:
                    type: number
                    format: float
                    example: 123.45
                  created_at:
                    type: string
                    example: 2024-01-01T12:00:00Z
                  ended_at:
                    type: string
                    example: 2024-01-31T12:00:00Z
                  code:
                    type: string
        "404":
          description: Product not found
    put:    
      summary: Update a product by ID
      tags: [Products]
      security:
        - BearerAuth: []        
      parameters:
        - name: product_type_id
          in: path
          required: true
          schema: { type: string, minimum: 1 }
        - name: product_id
          in: path
          required: true
          schema: { type: string, minimum: 1 }
      requestBody:      
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                new_product_name:
                  type: string
                  example: Updated Batch 
      responses:
        "200":
          description: Product updated successfully
        "400":
          description: Invalid input
        "401":
          description: Unauthorized
        "404":
          description: Product not found
    delete:
      summary: Delete a product by ID
      tags: [Products]
      security:
        - BearerAuth: []
      parameters:     
        - name: product_type_id  
          in: path
          required: true
          schema: { type: string, minimum: 1 }  
        - name: product_id
          in: path
          required: true        
          schema: { type: string, minimum: 1 }
      responses:
        "200":
          description: Product deleted successfully
          content:
            application/json: 
              schema:
                type: object
                properties:
                  status_message:
                    type: string
        "401":
          description: Unauthorized
        "404":
          description: Product not found 
  /products/{product_id}/steps/{stage_id}:
    get:
      summary: Get all steps order under a stage of a product
      tags: [Product Steps]
      security:
        - BearerAuth: []
      parameters:
        - name: product_id
          in: path
          required: true
          schema: { type: string, minimum: 1 }
        - name: stage_id
          in: path
          required: true
          schema: { type: string }   
      responses:
        "200":
          description: List of stage orders
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    step_id:
                      type: string
                    step_name:
                      type: string
                    tag_id:
                      type: string
                    sort_order:
                      type: integer
        "404":
          description: Product not found
  /products/{product_id}/steps:
    post:
      summary: Create a new step order under a stage of a product
      tags: [Product Steps]
      security:
        - BearerAuth: []
      parameters:
        - name: product_id  
          in: path
          required: true
          schema: { type: string, minimum: 1 }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                step_id:
                  type: string
                  example: STP1
                step_name: 
                  type: string
                  example: 原料採購
                tag_id:
                  type: string
                  example: TAG1
                sort_order:
                  type: integer
                  example: 1
                created_at:
                  type: string
                  example: 2024-01-01T12:00:00Z
      responses:
        "201":
          description: Stage order created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  step_id:
                    type: string
                  order:
                    type: integer
                  tag_id:
                    type: string
                  stage_id:
                    type: string
        "400":
          description: Invalid input
        "401":
          description: Unauthorized 
...